Added listeners annotation to BaseTest class. Also refactored user test suites and separated into positive and negative packages.
