--- src/test/java/com/ss/apitesting/user/negative/CreateListOfUsersNegativeTest.java
@@ -1,4 +1,0 @@
-package com.ss.apitesting.user.negative;
-
-public class CreateListOfUsersNegativeTest {
-}
