--- src/test/java/com/ss/apitesting/user/negative/LogoutUserNegativeTest.java
@@ -1,4 +1,0 @@
-package com.ss.apitesting.user.negative;
-
-public class LogoutUserNegativeTest {
-}
